fix: builder e2e specs
