+++ apps/builder-e2e/src/support/database/app.ts
@@ -1,10 +1,5 @@
 import type { IAppDTO } from '@codelab/frontend/abstract/core'
-import {
-  APP_PAGE_NAME,
-  APP_PAGE_SLUG,
-  DEFAULT_GET_SERVER_SIDE_PROPS,
-  ROOT_ELEMENT_NAME,
-} from '@codelab/frontend/abstract/core'
+import { APP_PAGE_NAME, APP_PAGE_SLUG } from '@codelab/frontend/abstract/core'
 import { createSlug } from '@codelab/frontend/shared/utils'
 import type { AppCreateInput } from '@codelab/shared/abstract/codegen'
 import { ITypeKind } from '@codelab/shared/abstract/core'
@@ -12,34 +7,10 @@ import { connectOwner } from '@codelab/shared/data'
 import { print } from 'graphql'
 import { CreateAppsDocument } from 'libs/frontend/domain/app/src/graphql/app.endpoints.graphql.gen'
 import { v4 } from 'uuid'
+import { createPageInput } from './page'
 
 export const createAppInput = (userId: string): AppCreateInput => {
   const appId = v4()
-  const pageId = v4()
-  const rootId = v4()
-
-  const providerPage = {
-    id: pageId,
-    name: APP_PAGE_NAME,
-    slug: `${appId}-${APP_PAGE_SLUG}`,
-    getServerSideProps: DEFAULT_GET_SERVER_SIDE_PROPS,
-    app: {
-      connect: { where: { node: { id: appId } } },
-    },
-    rootElement: {
-      create: {
-        node: {
-          id: rootId,
-          name: ROOT_ELEMENT_NAME,
-          slug: createSlug(ROOT_ELEMENT_NAME, pageId),
-        },
-      },
-    },
-    isProvider: true,
-    pageContainerElement: {
-      connect: { where: { node: { id: rootId } } },
-    },
-  }
 
   return {
     id: appId,
@@ -47,7 +18,18 @@ export const createAppInput = (userId: string): AppCreateInput => {
     slug: `test-${appId}`,
     owner: connectOwner(userId),
     pages: {
-      create: [{ node: providerPage }],
+      create: [
+        // create provider page
+        {
+          node: createPageInput(appId, {
+            name: APP_PAGE_NAME,
+            slug: createSlug(APP_PAGE_SLUG, appId),
+            isProvider: true,
+          }),
+        },
+        // create test page
+        { node: createPageInput(appId) },
+      ],
     },
     store: {
       create: {
@@ -73,10 +55,10 @@ export const createAppInput = (userId: string): AppCreateInput => {
   }
 }
 
-export const createApp = (userId: string, input?: AppCreateInput) =>
+export const createApp = (input: AppCreateInput) =>
   cy
     .graphqlRequest({
       query: print(CreateAppsDocument),
-      variables: { input: input || createAppInput(userId) },
+      variables: { input },
     })
     .then((result) => result.body.data?.createApps.apps as Array<IAppDTO>)
