+++ apps/builder-e2e/src/support/database/page.ts
@@ -1,45 +1,42 @@
-import type { IPageDTO } from '@codelab/frontend/abstract/core'
-import { ROOT_ELEMENT_NAME } from '@codelab/frontend/abstract/core'
+import {
+  DEFAULT_GET_SERVER_SIDE_PROPS,
+  ROOT_ELEMENT_NAME,
+} from '@codelab/frontend/abstract/core'
 import { createSlug } from '@codelab/frontend/shared/utils'
-import type {
-  PageCreateInput,
-  PageWhere,
-} from '@codelab/shared/abstract/codegen'
-import { print } from 'graphql'
-import { GetPagesDocument } from 'libs/frontend/domain/page/src/graphql'
+import type { PageCreateInput } from '@codelab/shared/abstract/codegen'
+import merge from 'lodash/merge'
 import { v4 } from 'uuid'
 
-export const getPages = (input: PageWhere) =>
-  cy
-    .graphqlRequest({
-      query: print(GetPagesDocument),
-      variables: { input },
-    })
-    .then((result) => result.body.data?.pages as Array<IPageDTO>)
-
-export const createPageInput = (appId: string): PageCreateInput => {
+export const createPageInput = (
+  appId: string,
+  input: Partial<PageCreateInput> = {},
+): PageCreateInput => {
   const id = v4()
+  const rootId = v4()
+  const name = 'Test Page'
 
-  return {
-    id: id,
-    name: 'Test Page',
-    rootElement: {
-      create: {
-        node: {
-          id: v4(),
-          name: ROOT_ELEMENT_NAME,
-          slug: createSlug(ROOT_ELEMENT_NAME, id),
+  return merge(
+    {
+      id: id,
+      name,
+      getServerSideProps: DEFAULT_GET_SERVER_SIDE_PROPS,
+      rootElement: {
+        create: {
+          node: {
+            id: rootId,
+            name: ROOT_ELEMENT_NAME,
+            slug: createSlug(ROOT_ELEMENT_NAME, id),
+          },
         },
       },
+      app: {
+        connect: { where: { node: { id: appId } } },
+      },
+      slug: createSlug(name, appId),
+      pageContainerElement: {
+        connect: { where: { node: { id: rootId } } },
+      },
     },
-    slug: createSlug('Test Page', appId),
-  }
+    input,
+  )
 }
-
-export const createPage = (input: Partial<PageCreateInput>) =>
-  cy
-    .graphqlRequest({
-      query: print(CreatePagesDocument),
-      variables: { input: { ...createPageInput(), ...input } },
-    })
-    .then((result) => result.body.data?.createPages.pages as Array<IPageDTO>)
