--- apps/builder-e2e/src/support/helpers/createPageFromScratch.ts
@@ -1,10 +1,0 @@
-export const createPageFromScratch = () =>
-  cy.createApp().then((apps) => {
-    cy.createPage({
-      app: { connect: { where: { node: { id: apps[0]?.id } } } },
-    }).then((page) => ({
-      appId: apps[0]?.id,
-      pageId: page[0]?.id,
-      rootElementId: page[0].rootElement.id,
-    }))
-  })
